+++ src/mpservice/util.py
@@ -1,3 +1,4 @@
+from __future__ import annotations
 import errno
 import functools
 import inspect
@@ -13,7 +14,7 @@ import threading
 import traceback
 from multiprocessing.util import Finalize
 from types import TracebackType
-from typing import Optional, Union
+from typing import Optional
 
 
 MAX_THREADS = min(32, (os.cpu_count() or 1) + 4)
@@ -168,7 +169,7 @@ class RemoteException:
     #
     # Also check out `sys.excepthook`.
 
-    def __init__(self, exc: Exception, tb: Optional[Union[TracebackType, str]] = None):
+    def __init__(self, exc: Exception, tb: Optional[TracebackType | str] = None):
         if isinstance(tb, str):
             pass
         elif isinstance(tb, TracebackType):
@@ -228,6 +229,9 @@ class Thread(threading.Thread):
         self._loud_exception_ = loud_exception
 
     def run(self):
+        """
+        This method represents the thread's activity.
+        """
         try:
             if self._target is not None:
                 self._result_ = self._target(*self._args, **self._kwargs)
@@ -240,7 +244,7 @@ class Thread(threading.Thread):
             # an argument that has a member that points to the thread.
             del self._target, self._args, self._kwargs
 
-    def done(self):
+    def done(self) -> bool:
         if self.is_alive():
             return False
         return self._started.is_set()
@@ -323,7 +327,7 @@ class SpawnProcess(multiprocessing.context.SpawnProcess):
             result_and_error.send(None)
             result_and_error.send(None)
 
-    def done(self):
+    def done(self) -> bool:
         return self.exitcode is not None
 
     def result(self, timeout=None):